fix(desktop-v3): sanitize reflection question output too
The reflection generator has the same over-production leak as the
briefing — Phi-3-q4 writes the question, then appends a 'Solution N:'
follow-up after a blank line, which got stored as the question (visible
as contaminated reflection rows).

Move first_paragraph to the shared prompts module (pub(crate)) and apply
it in both briefing generation and reflection's build_question. Briefing
behavior is unchanged (same fn, now shared). Reflection questions are now
truncated to the question itself. Unit-tested in both modules.
